% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{geom_arrow}
\alias{geom_arrow}
\title{geom_arrow}
\usage{
geom_arrow(mapping = NULL, data = NULL, stat = "arrow",
  position = "identity", ..., start = 0, direction = 1,
  min.mag = 0, skip = 0, skip.x = skip, skip.y = skip,
  arrow.angle = 15, arrow.length = 0.5, arrow.ends = "last",
  arrow.type = "closed", arrow = grid::arrow(arrow.angle,
  unit(arrow.length, "lines"), ends = arrow.ends, type = arrow.type),
  lineend = "butt", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{inherit.aes}{}
}
\description{
geom_arrow
}
\examples{

library(tibble)
geo <- tibble(lon = 1:10, lat = 1:10, mag = 1:10, angle = 1:10)

# scale_mag_continuous <- scale_mag

ggplot(geo, aes(lon, lat)) +
    geom_arrow(aes(mag = mag, angle = angle))

}
